new organization of the code in remove_gravity.py
added rotationMatrix.py which contains a function that only computes the rotation matrix R
