+++ smartwatch/src/rotationMatrix.py
@@ -1,0 +1,26 @@
+#!/usr/bin/env python
+
+import numpy as np
+import math
+#Compute rotation matrix R, starting from the angles given in euler rapresentation
+
+def eulerAnglesToRotationMatrix(angles):  # angles [roll, pitch, yaw]
+
+    R_x = np.array([[1,         0,                  0],
+                    [0,         math.cos(angles[0]),   math.sin(angles[0])],
+                    [0,         -math.sin(angles[0]),  math.cos(angles[0])]
+                    ])
+
+    R_y = np.array([[math.cos(angles[1]),    0,      -math.sin(angles[1])],
+                    [0,                      1,      0],
+                    [math.sin(angles[1]),    0,      math.cos(angles[1])]
+                    ])
+
+    R_z = np.array([[math.cos(angles[2]),      math.sin(angles[2]),     0],
+                    [-math.sin(angles[2]),     math.cos(angles[2]),     0],
+                    [0,                        0,                       1]
+                    ])
+
+    R = np.dot(R_x, np.dot(R_y, R_z))
+
+    return R
